+++ common/src/main/java/com/tc/net/core/TCChannelWriter.java
@@ -28,5 +28,5 @@ import java.io.IOException;
  * @author teck
  */
 interface TCChannelWriter {
-  int doWrite() throws IOException;
+  long doWrite() throws IOException;
 }
